requirements.txt
opencv_python==4.8.0.74
Pillow==9.1.0
Pillow==10.0.0
torch==1.8.1+cu111
ultralytics==8.0.128
